# users-service/swagger/users-service-swagger.yaml
openapi: 3.0.3
info:
  title: Raspstore User Service
  description: Microservice REST specification of users-service endpoints.
  contact:
    email: soon@todo.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
servers:
  - url: http://localhost:9090/users-service/api/v1
...
paths:
  /users:
    post:
      tags:
        - users
      summary: Create user
      description: Create new user
      operationId: createUser
      requestBody:
        $ref: '#/components/requestBodies/CreateUserRequest'
      responses:
        '201':
          $ref: '#/components/responses/SuccessCreateUserResponse'
        '400':
          $ref: '#/components/responses/BadRequestCreateUserResponse'
        '422':
          description: "Unprocessable Entity"
        '500':
          $ref: '#/components/responses/InternalServerErrorResponse'
    get:
      tags:
        - users
      summary: Get All Users
      description: Get all users based on parameters provided
      operationId: getUsers
      parameters:
        - $ref: '#/components/parameters/PageQueryParameter'
        - $ref: '#/components/parameters/SizeQueryParameter'
        - $ref: '#/components/parameters/UsernameQueryParameter'
        - $ref: '#/components/parameters/EmailQueryParameter'
        - $ref: '#/components/parameters/EnableQueryParameter'
      responses:
        '200':
          $ref: '#/components/responses/SuccessUsersListResponse'
...
components:
  schemas:
    UserRepresentation:
      type: object
      properties:
        userId:
          type: string
          example: 507f1f77bcf86cd799439011
        username:
          type: string
          example: johndoe
        email:
          type: string
          example: john.doe@dummy.com
        phoneNumber:
          type: string
          example: +1111555333
        createdAt:
          type: string
          format: date
          pattern: YYYY-MM-DD HH:mm:ss
          example: '2022-10-25 19:43:12'
        updatedAt:
          type: string
          format: date
          pattern: YYYY-MM-DD HH:mm:ss
          example: '2022-10-25 19:43:12'
    CreateUserRepresentation:
      type: object
      properties:
        username:
          type: string
          example: johndoe
        email:
          type: string
          example: john.doe@dummy.com
        password:
          type: string
          example: super-secret-password
        phoneNumber:
          type: string
          example: +1111555333
...
    PageRepresentation:
      type: object
      properties:
        size:
          type: integer
          example: 10
        totalElements:
          type: integer
          example: 10
        page:
          type: integer
          example: 1
        next:
          type: string
          example: 'http://localhost:9000/users?page=2&size=10'
...
    ApiErrorException:
      type: object
      properties:
        code:
          type: string
          example: 'ERR010'
        message:
          type: string
          example: 'Field email must not be null or empty'
        traceId:
          type: string
          example: 'df438c36-9aa4-11ed-a8fc-0242ac120002'
  requestBodies:
    CreateUserRequest:
      description: User request object required to create a new user
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/CreateUserRepresentation'
...
  parameters:
    UsernameQueryParameter:
      name: username
      in: query
      schema:
        type: string
        example: johndoe
    EmailQueryParameter:
      name: email
      in: query
      schema:
        type: string
        example: john.doe@mail.com
    EnableQueryParameter:
      name: enabled
      in: query
      schema:
        type: boolean
        example: true
    PageQueryParameter:
      name: page
      required: true
      in: query
      schema:
        type: integer
        example: 1
    SizeQueryParameter:
      name: size
      required: true
      in: query
      schema:
        type: integer
        example: 1
...
  responses:
    BadRequestCreateUserResponse:
      description: payload invalid or malformed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorException'
    SuccessCreateUserResponse:
      description: successful operation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/UserRepresentation'
...
    SuccessUsersListResponse:
      description: List of user object
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/PageRepresentation'
              - type: object
                properties:
                  content:
                    type: array
                    items:
                      $ref: '#/components/schemas/UserRepresentation'
    InternalServerErrorResponse:
      description: internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorException'
          example:
            code: "ERR500"
            message: "An expected error occured"
            traceId: "c1611200-9aa5-11ed-a8fc-0242ac120002"
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []